Team6_IoTProject_FireDetection/Team6_PythonCode/requirements.txt
numpy==1.22.4
opencv-python==4.5.5.64
paho-mqtt==1.6.1
pyserial==3.5
